+++ packages/gatsby/cache-dir/app.js
@@ -22,11 +22,14 @@ import "./blank.css"
 // Enable fast-refresh for virtual sync-requires, gatsby-browser & navigation
 // To ensure that our <Root /> component can hot reload in case anything below doesn't
 // satisfy fast-refresh constraints
-module.hot.accept([
-  `$virtual/async-requires`,
-  `./api-runner-browser`,
-  `./navigation`,
-])
+module.hot.accept(
+  [`$virtual/async-requires`, `./api-runner-browser`, `./navigation`],
+  () => {
+    // asyncRequires should be automatically updated here (due to ESM import and webpack HMR spec),
+    // but loader doesn't know that and needs to be manually nudged
+    loader.updateAsyncRequires(asyncRequires)
+  }
+)
 
 window.___emitter = emitter
 
@@ -131,8 +134,8 @@ apiRunnerAsync(`onClientEntry`).then(() => {
   // render to avoid React complaining about hydration mis-matches.
   let defaultRenderer = ReactDOM.render
   if (focusEl && focusEl.children.length) {
-    if (ReactDOM.createRoot) {
-      defaultRenderer = ReactDOM.createRoot
+    if (ReactDOM.hydrateRoot) {
+      defaultRenderer = ReactDOM.hydrateRoot
     } else {
       defaultRenderer = ReactDOM.hydrate
     }
@@ -190,7 +193,7 @@ apiRunnerAsync(`onClientEntry`).then(() => {
         )
         document.body.append(indicatorMountElement)
 
-        if (renderer === ReactDOM.createRoot) {
+        if (renderer === ReactDOM.hydrateRoot) {
           renderer(indicatorMountElement).render(
             <LoadingIndicatorEventHandler />
           )
@@ -222,10 +225,8 @@ apiRunnerAsync(`onClientEntry`).then(() => {
         dismissLoadingIndicator()
       }
 
-      if (renderer === ReactDOM.createRoot) {
-        renderer(rootElement, {
-          hydrate: true,
-        }).render(<App />)
+      if (renderer === ReactDOM.hydrateRoot) {
+        renderer(rootElement, <App />)
       } else {
         renderer(<App />, rootElement)
       }